No longer marking fd as INVALID_FD for removal
Marking the fd of the pollfd struct as INVALID_FD (-1) does not work in our
setup. When the fd in the pollfd array/vector is changed into INVALID_FD
it is no longer possible to find correlating pollable (in _pollables
map).

For now I'm using a boolean to mark a pollable for removal.

All pollables are checked for timeout and removal in every poll loop.