Refactored receiving header/body and responding
1. Try/Catch block has been moved from receiveHeader() to process() and
now encapsulates receiveHeader(), receiveBody() and respond().

2. receiveHeader() no longer initializes a response (in case of a
'Expect: 100-continue'), but instead directs the process to respond(),
where all responses (bar the error pages) are made.

3. respond() will initialize a '100-continue' response and direct the
process to receiveBody (_state = BODY) on next ClientFD::pollin() iteration.

4. ClientFD::pollout() will reset the ClientFD in order to accept new
incoming activity, unless:
	a) 'Connection: close' was present in request;
	b) The response sent was a '100-continue', thus only reset the
	byte counters;

The order of requests/responses in case of an initial request with
'Expect: 100-continue' is now as follows:
< Request ... Expect: 100-continue;
> Response ... 100 Continue;
< Receive body;
> Final response;
< Create File.

In short: after sending the final response the file is created.

Small note on ClientFD::pollout(): for now the _state is checked for
ERROR, while the resulting behavior doesn't differ from 'normal'
responses, it has to be checked before getting to the '100-continue'
response. I'll have to look at this again.