Refactored pollables map into vector
The main selling point for using a map at the start was fast lookup
using a file descriptor.

Now, later in the project, the lookup was barely used. It rather was
cumbersome having to switch between the vector (of pollfd structs) and
map (of pollables). Extra lines of code was the result.

With that in mind the pollables are now stored inside a vector. The
index of a pollfd struct inside _pfds correlates with the index of the
related pollable inside _pollables.

After removing a pollable (swapping the last element with the element to
be removed), index is updated.