Addition of updateTick and getTick in pollables
On every activity of the pollables (POLLIN/POLLOUT) updateTick() will
call time() and register the time in '_tick' attribute.

To get the time of the most recent activity use getTick().

The time() call uses a time_t struct. To make sense of the value stored
in this struct, other functions need to be used.

(such as: https://en.cppreference.com/w/cpp/chrono/c/localtime)