Ro/updatetick (#67)
* Moved the try catch block from receiveheader to process. After sending an error page, the connection should accept new requests. With 'Expect: 100-continue', after sending an error (405), it continues with receiving a body.

* Refactored receiving header/body and responding

1. Try/Catch block has been moved from receiveHeader() to process() and
now encapsulates receiveHeader(), receiveBody() and respond().

2. receiveHeader() no longer initializes a response (in case of a
'Expect: 100-continue'), but instead directs the process to respond(),
where all responses (bar the error pages) are made.

3. respond() will initialize a '100-continue' response and direct the
process to receiveBody (_state = BODY) on next ClientFD::pollin() iteration.

4. ClientFD::pollout() will reset the ClientFD in order to accept new
incoming activity, unless:
	a) 'Connection: close' was present in request;
	b) The response sent was a '100-continue', thus only reset the
	byte counters;

The order of requests/responses in case of an initial request with
'Expect: 100-continue' is now as follows:
< Request ... Expect: 100-continue;
> Response ... 100 Continue;
< Receive body;
> Final response;
< Create File.

In short: after sending the final response the file is created.

Small note on ClientFD::pollout(): for now the _state is checked for
ERROR, while the resulting behavior doesn't differ from 'normal'
responses, it has to be checked before getting to the '100-continue'
response. I'll have to look at this again.

* Made some new comments

* when client send a SIGINT, shutdown ClientFD

* Return values of write/read/send/recv

Now handles return values of '0' and '-1'.

In the cases of '0':
	- recv(): the client has closed its side of the connection; close
	pollable.

	- send(): send could return 0 when len of size 0 is passed as
	argument (send 0 bytes basicly). This is caught in the code
	block after, because this means all data is sent.

	- read(): read the file; close pollable; send response with file
	content.

	- write(): same as send(); the file is succesfully created; send
	a response.

In the cases of '-1':
	- recv(): poll signals there is data to be read, the subsequent
	recv() call returns -1, which is most likely to be EAGAIN or
	EWOULDBLOCK, no data to be read e.g. waiting for data would block.
	Therefor send an internal server error response.

	- send(): an error on send; no point trying to send again
	because this could create an infinite loop; close the pollable.

	- read(): send internal server error response.

	- write(): send internal server error response.

* Receiving chunked, removed duplicate str.find(rn), removed lot of comments, expanded comment about 100-continue response

* Error handling on read/write/receive/send

When write, read, or receive returns an error (-1): send an internal error
response to client and close connection.

When send returns an error, close connection.

When receive returns a shutdown (0), close connection.

* Minor spelling correction

* updateTick

Previously the ClientFD could time out while its child (the FileFD) was still reading or writing. Resulting in an error eventually.
The FileFD will now also update the _tick of ClientFD on read or write.

* use system defined max connections for backlog in listen()

* Addition of updateTick and getTick in pollables

On every activity of the pollables (POLLIN/POLLOUT) updateTick() will
call time() and register the time in '_tick' attribute.

To get the time of the most recent activity use getTick().

The time() call uses a time_t struct. To make sense of the value stored
in this struct, other functions need to be used.

(such as: https://en.cppreference.com/w/cpp/chrono/c/localtime)

* Replaced time() call with updateTick() in the constructors of the pollables.

Co-authored-by: Ronald Heuts <rheuts@f0r6s1.codam.nl>
Co-authored-by: Ronald Heuts <rheuts@f0r3s26.codam.nl>
Co-authored-by: Ronald Heuts <rheuts@f0r1s6.codam.nl>
Co-authored-by: Ronald Heuts <rheuts@f0r1s1.codam.nl>