Ro/factory (#71)
* PollableFactory

Replacement of the Server::addPollable.

Now no longer uses switch case (because that's icky?).

The factory returns the specified pollable type (ServerFD, ClientFD
or FileFD) and sets the event to poll for.

usage:
PollableFactory::pf.createPollable(server, file descriptor, pollable
type, poll event)

misc modifications:
- shortened getFileDescriptor() to getFD()
- replaced the old Server::addPollable with the new Factory.

Vamos!

* Refactored PollableFactory to adhere to Singleton customs.

* Moved the try catch block from receiveheader to process. After sending an error page, the connection should accept new requests. With 'Expect: 100-continue', after sending an error (405), it continues with receiving a body.

* Refactored receiving header/body and responding

1. Try/Catch block has been moved from receiveHeader() to process() and
now encapsulates receiveHeader(), receiveBody() and respond().

2. receiveHeader() no longer initializes a response (in case of a
'Expect: 100-continue'), but instead directs the process to respond(),
where all responses (bar the error pages) are made.

3. respond() will initialize a '100-continue' response and direct the
process to receiveBody (_state = BODY) on next ClientFD::pollin() iteration.

4. ClientFD::pollout() will reset the ClientFD in order to accept new
incoming activity, unless:
	a) 'Connection: close' was present in request;
	b) The response sent was a '100-continue', thus only reset the
	byte counters;

The order of requests/responses in case of an initial request with
'Expect: 100-continue' is now as follows:
< Request ... Expect: 100-continue;
> Response ... 100 Continue;
< Receive body;
> Final response;
< Create File.

In short: after sending the final response the file is created.

Small note on ClientFD::pollout(): for now the _state is checked for
ERROR, while the resulting behavior doesn't differ from 'normal'
responses, it has to be checked before getting to the '100-continue'
response. I'll have to look at this again.

* Made some new comments

* when client send a SIGINT, shutdown ClientFD

* Return values of write/read/send/recv

Now handles return values of '0' and '-1'.

In the cases of '0':
	- recv(): the client has closed its side of the connection; close
	pollable.

	- send(): send could return 0 when len of size 0 is passed as
	argument (send 0 bytes basicly). This is caught in the code
	block after, because this means all data is sent.

	- read(): read the file; close pollable; send response with file
	content.

	- write(): same as send(); the file is succesfully created; send
	a response.

In the cases of '-1':
	- recv(): poll signals there is data to be read, the subsequent
	recv() call returns -1, which is most likely to be EAGAIN or
	EWOULDBLOCK, no data to be read e.g. waiting for data would block.
	Therefor send an internal server error response.

	- send(): an error on send; no point trying to send again
	because this could create an infinite loop; close the pollable.

	- read(): send internal server error response.

	- write(): send internal server error response.

* Refactored pollables map into vector

The main selling point for using a map at the start was fast lookup
using a file descriptor.

Now, later in the project, the lookup was barely used. It rather was
cumbersome having to switch between the vector (of pollfd structs) and
map (of pollables). Extra lines of code was the result.

With that in mind the pollables are now stored inside a vector. The
index of a pollfd struct inside _pfds correlates with the index of the
related pollable inside _pollables.

* Refactored pollables map into vector

The main selling point for using a map at the start was fast lookup
using a file descriptor.

Now, later in the project, the lookup was barely used. It rather was
cumbersome having to switch between the vector (of pollfd structs) and
map (of pollables). Extra lines of code was the result.

With that in mind the pollables are now stored inside a vector. The
index of a pollfd struct inside _pfds correlates with the index of the
related pollable inside _pollables.

After removing a pollable (swapping the last element with the element to
be removed), index is updated.

* On POLLHUP, try to receive remaining bytes, else close the socket

* enums are now: SERVERPOLL, CLIENTPOLL, FILEPOLL and don't share the same name as the classes

* pre-merge

* Pulled updated httprequest

* now uses system definedmax connections

* createPollable now uses PollableType instead of...

...int type for creating a pollable.

Co-authored-by: Ronald Heuts <rheuts@f0r1s6.codam.nl>
Co-authored-by: Ronald Heuts <rheuts@f0r3s26.codam.nl>
Co-authored-by: Ronald Heuts <rheuts@f0r6s1.codam.nl>
Co-authored-by: Ronald Heuts <rheuts@f0r1s1.codam.nl>